Feat: Separete js script in each file
