+++ index.html
@@ -12,7 +12,7 @@
       crossorigin="anonymous"
       referrerpolicy="no-referrer"
     />
-    <script type="module" src="./main.js"></script>
+    <script type="module" src="./js/main.js"></script>
   </head>
   <body>
     <section class="home">
