+++ js/main.js
@@ -1,0 +1,3 @@
+import "./menuToggle.js";
+import "./destinationImageSlider.js";
+import "./insideImageSlider.js";
