+++ package.json
@@ -1,0 +1,5 @@
+{
+  "dependencies": {
+    "node-sass": "^9.0.0"
+  }
+}
